+++ pelutils/ds/__init__.py
@@ -57,55 +57,14 @@ def unique(
     return tuple(ret) if len(ret) > 1 else ret[0]
 
 def no_grad(fun: Callable) -> Callable:
-    """
-    Decorator for running functions without pytorch tracking gradients, e.g.
+    """ Decorator for running functions without pytorch tracking gradients, e.g.
     ```
     @no_grad
     def feed_forward(x):
         return net(x)
-    ```
-    """
+    ``` """
     functools.wraps(fun)
     def wrapper(*args, **kwargs):
         with torch.no_grad():
             return fun(*args, **kwargs)
     return wrapper
-
-
-class BatchFeedForward:
-    """
-    This class handles feedforwarding large batches that would otherwise cause memory overflow
-    It works by splitting it into smaller batches, if it encounters a memory error
-    Does not track gradients
-    Notice that while this works for batches of varying sizes, this is not recommended usage, as it can be inefficient
-    """
-
-    def __init__(self, net: torch.nn.Module):
-        self.net = net
-        self.increase_factor = 2
-        self.batches = 1
-
-    @no_grad
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        while True:
-            try:
-                output_parts = [self.net(x[slice_]) for slice_ in self._get_slices(x)]
-                output = torch.cat(output_parts)
-                break
-            # Usually caused by running out of vram. If not, the error is still raised, else batch size is reduced
-            except RuntimeError as e:
-                if "alloc" not in str(e):
-                    raise e
-                self._more_batches()
-        return output
-
-    def __call__(self, x: torch.Tensor) -> torch.Tensor:
-        return self.forward(x)
-
-    def _more_batches(self):
-        self.batches *= self.increase_factor
-
-    def _get_slices(self, x: torch.Tensor) -> Generator:
-        slice_size = len(x) // self.batches + 1
-        # Final slice may have overflow, however this is simply ignored when indexing
-        return (slice(i*slice_size, (i+1)*slice_size) for i in range(self.batches))