+++ CHANGELOG.md
@@ -38,6 +38,9 @@
 - Made C functions to work on most platforms
 - Added `is_windows` function
 - Removed `TickTock.profile_iter`
+- Added `SimplePool` for testing with `multiprocessing.Pool`
+- Renamed `MainTest` to `UnitTestCollection`
+- Removed `BatchFeedForward`
 
 ### Bug fixes